app/kpi-analytics/kpi-serving/requirements.txt
fastapi>=0.109.0,<1
uvicorn[standard]>=0.27.0,<1
google-cloud-bigquery>=3.17.0,<4
pydantic>=2.5.0,<3
pydantic-settings>=2.1.0,<3
